+++ NEWS.md
@@ -6,6 +6,7 @@ No longer force render data as JSON if using an HTML template.
 - Added `jobj` function to serialise objects to JSON in `render`.
 - Allow passing `host` and `port` to `start()` method.
 - Added `host` and `port` active bindings.
+- Move internal `is_running` field to private.
 
 # ambiorix 2.0.0